requirements.txt
alembic==1.8.0
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0
flake8==4.0.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.2.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.21.6
openpyxl==3.0.10
pandas==1.1.5
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
